refactor: consolidate event-related enums into single file
