--- src/common/enums/event-level.enum.ts
@@ -1,7 +1,0 @@
-export enum EventLevel {
-  INFO = 'INFO',
-  WARNING = 'WARNING',
-  FAULT = 'FAULT',
-  CRITICAL = 'CRITICAL',
-  ERROR = 'ERROR',
-}
