--- src/common/enums/event-tag.enum.ts
@@ -1,6 +1,0 @@
-export enum EventTag {
-  PQC_GATEWAY = 'PQC-GATEWAY',
-  SYSTEM = 'SYSTEM',
-  DEVICE = 'DEVICE',
-  USER = 'USER',
-}
