--- src/common/enums/event-type.enum.ts
@@ -1,5 +1,0 @@
-export enum EventType {
-  DEVICE_REGISTRATION = 'DEVICE-REGISTRATION',
-  DEVICE_AUTHENTICATION = 'DEVICE-AUTHENTICATION',
-  PQC_GATEWAY_ALARM = 'PQC-GATEWAY-ALARM',
-}
