+++ src/common/enums/event.enum.ts
@@ -1,0 +1,20 @@
+export enum EventLevel {
+  INFO = 'INFO',
+  WARNING = 'WARNING',
+  FAULT = 'FAULT',
+  CRITICAL = 'CRITICAL',
+  ERROR = 'ERROR',
+}
+
+export enum EventTag {
+  PQC_GATEWAY = 'PQC-GATEWAY',
+  SYSTEM = 'SYSTEM',
+  DEVICE = 'DEVICE',
+  USER = 'USER',
+}
+
+export enum EventType {
+  DEVICE_REGISTRATION = 'DEVICE-REGISTRATION',
+  DEVICE_AUTHENTICATION = 'DEVICE-AUTHENTICATION',
+  PQC_GATEWAY_ALARM = 'PQC-GATEWAY-ALARM',
+}
