+++ src/common/enums/index.ts
@@ -1,7 +1,5 @@
 export * from './error-codes.enum';
 export * from './flow-control-level.enum';
 export * from './sort.enum';
-export * from './event-level.enum';
-export * from './event-type.enum';
-export * from './event-tag.enum';
 export * from './account-role.enum';
+export * from './event.enum';
